Added x86 support

+++ src/instructions/segmentation.rs
@@ -24,7 +24,7 @@ pub unsafe fn set_cs(sel: SegmentSelector) {
             "push {tmp}",
             "retfq",
             "1:",
-            sel = in(reg) u64::from(sel.0),
+            sel = in(reg) usize::from(sel.0),
             tmp = lateout(reg) _,
         );
     }
@@ -145,6 +145,7 @@ pub fn cs() -> SegmentSelector {
     }
 }
 
+#[cfg(target_arch = "x86_64")]
 /// Writes the FS segment base address
 ///
 /// ## Safety
@@ -171,6 +172,7 @@ pub unsafe fn wrfsbase(val: u64) {
     inner(val)
 }
 
+#[cfg(target_arch = "x86_64")]
 /// Reads the FS segment base address
 ///
 /// ## Safety
@@ -195,6 +197,7 @@ pub unsafe fn rdfsbase() -> u64 {
     inner()
 }
 
+#[cfg(target_arch = "x86_64")]
 /// Writes the GS segment base address
 ///
 /// ## Safety
@@ -220,6 +223,7 @@ pub unsafe fn wrgsbase(val: u64) {
     inner(val)
 }
 
+#[cfg(target_arch = "x86_64")]
 /// Reads the GS segment base address
 ///
 /// ## Safety
